feat: extract Header components to src/components/layout/Header.tsx
Move Header(), PhaseBadge, StatusDot, ConnectionIndicator, ElapsedTimer,
QACoverageBadge and their shared constants (phaseBarColors, statusColors)
from AppView.tsx into a dedicated layout component file.

AppView.tsx imports these from the new file; all 129 tests pass and
type-check is clean.

Aloop-Agent: build
Aloop-Iteration: 2
Aloop-Session: orchestrator-20260321-172932-issue-157-20260323-074811
